added rating list tiles
