+++ lib/ui/widgets/account/components/settings_card_widget.dart
@@ -28,7 +28,7 @@ class _SettingsCardWidgetState extends State<SettingsCardWidget> {
       },
       title: Text(
         'Change color theme',
-        style: Theme.of(context).textTheme.headlineMedium,
+        style: Theme.of(context).textTheme.displayMedium,
       ),
     );
   }
